chore: some more file renames to camelCase
