+++ docs/source/api.js
@@ -36,9 +36,9 @@ const files = [
   'lib/options/SchemaNumberOptions.js',
   'lib/options/SchemaObjectIdOptions.js',
   'lib/options/SchemaStringOptions.js',
-  'lib/types/DocumentArray/methods/index.js',
+  'lib/types/documentArray/methods/index.js',
   'lib/types/subdocument.js',
-  'lib/types/ArraySubdocument.js',
+  'lib/types/arraySubdocument.js',
   'lib/types/buffer.js',
   'lib/types/decimal128.js',
   'lib/types/map.js',
