+++ lib/document.js
@@ -3533,7 +3533,7 @@ Document.prototype.$__setSchema = function(schema) {
  */
 
 Document.prototype.$__getArrayPathsToValidate = function() {
-  DocumentArray || (DocumentArray = require('./types/DocumentArray'));
+  DocumentArray || (DocumentArray = require('./types/documentArray'));
 
   // validate all document arrays.
   return this.$__.activePaths
@@ -3562,7 +3562,7 @@ Document.prototype.$__getArrayPathsToValidate = function() {
  */
 
 Document.prototype.$getAllSubdocs = function() {
-  DocumentArray || (DocumentArray = require('./types/DocumentArray'));
+  DocumentArray || (DocumentArray = require('./types/documentArray'));
   Embedded = Embedded || require('./types/ArraySubdocument');
 
   function docReducer(doc, seed, path) {
