+++ lib/options/virtualOptions.js
