+++ lib/schema.js
@@ -9,7 +9,7 @@ const Kareem = require('kareem');
 const MongooseError = require('./error/mongooseError');
 const SchemaType = require('./schemaType');
 const SchemaTypeOptions = require('./options/SchemaTypeOptions');
-const VirtualOptions = require('./options/VirtualOptions');
+const VirtualOptions = require('./options/virtualOptions');
 const VirtualType = require('./virtualType');
 const addAutoId = require('./helpers/schema/addAutoId');
 const clone = require('./helpers/clone');
