+++ lib/schema/documentArray.js
@@ -5,7 +5,7 @@
  */
 
 const CastError = require('../error/cast');
-const DocumentArrayElement = require('./DocumentArrayElement');
+const DocumentArrayElement = require('./documentArrayElement');
 const EventEmitter = require('events').EventEmitter;
 const SchemaArray = require('./array');
 const SchemaDocumentArrayOptions =
@@ -221,7 +221,7 @@ SchemaDocumentArray.prototype.discriminator = function(name, schema, options) {
 
 SchemaDocumentArray.prototype.doValidate = function(array, fn, scope, options) {
   // lazy load
-  MongooseDocumentArray || (MongooseDocumentArray = require('../types/DocumentArray'));
+  MongooseDocumentArray || (MongooseDocumentArray = require('../types/documentArray'));
 
   const _this = this;
   try {
@@ -358,7 +358,7 @@ SchemaDocumentArray.prototype.getDefault = function(scope, init, options) {
   }
 
   // lazy load
-  MongooseDocumentArray || (MongooseDocumentArray = require('../types/DocumentArray'));
+  MongooseDocumentArray || (MongooseDocumentArray = require('../types/documentArray'));
 
   if (!Array.isArray(ret)) {
     ret = [ret];
@@ -397,7 +397,7 @@ const initDocumentOptions = Object.freeze({ skipId: false, willInit: true });
 
 SchemaDocumentArray.prototype.cast = function(value, doc, init, prev, options) {
   // lazy load
-  MongooseDocumentArray || (MongooseDocumentArray = require('../types/DocumentArray'));
+  MongooseDocumentArray || (MongooseDocumentArray = require('../types/documentArray'));
 
   // Skip casting if `value` is the same as the previous value, no need to cast. See gh-9266
   if (value != null && value[arrayPathSymbol] != null && value === prev) {
