+++ lib/schema/documentArrayElement.js
