+++ lib/types/documentArray/index.js
