+++ lib/types/documentArray/methods/index.js
