+++ lib/types/index.js
@@ -11,7 +11,7 @@ exports.Buffer = require('./buffer');
 exports.Document = // @deprecate
 exports.Embedded = require('./arraySubdocument');
 
-exports.DocumentArray = require('./DocumentArray');
+exports.DocumentArray = require('./documentArray');
 exports.Decimal128 = require('./decimal128');
 exports.ObjectId = require('./objectid');
