+++ lib/utils.js
@@ -13,7 +13,7 @@ const clone = require('./helpers/clone');
 const immediate = require('./helpers/immediate');
 const isObject = require('./helpers/isObject');
 const isMongooseArray = require('./types/array/isMongooseArray');
-const isMongooseDocumentArray = require('./types/DocumentArray/isMongooseDocumentArray');
+const isMongooseDocumentArray = require('./types/documentArray/isMongooseDocumentArray');
 const isBsonType = require('./helpers/isBsonType');
 const getFunctionName = require('./helpers/getFunctionName');
 const isMongooseObject = require('./helpers/isMongooseObject');
